MEL-25_p1_marketing_intel/web-scraper/requirements.txt
requests==2.31.0
psycopg2-binary==2.9.7
redis==5.0.1
beautifulsoup4==4.12.2
selenium==4.15.2
lxml==4.9.3
pandas==2.1.3
python-dotenv==1.0.0
aiohttp==3.9.1
fake-useragent==1.4.0
fastapi==0.104.1
uvicorn==0.24.0 